final add tests
