+++ scrapper/src/test/java/edu/java/scrapper/domain/service/AbstractTgChatServiceTest.java
@@ -1,25 +1,36 @@
-package edu.java.scrapper.domain.service.jdbc;
+package edu.java.scrapper.domain.service;
 
 import edu.java.domain.service.TgChatService;
 import edu.java.exceptions.tracker_exceptions.ChatNotFoundException;
 import edu.java.exceptions.tracker_exceptions.ChatReregisteringException;
 import edu.java.scrapper.IntegrationTest;
-import edu.java.scrapper.domain.service.AbstractTgChatServiceTest;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.Rollback;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.transaction.annotation.Transactional;
 import org.testcontainers.junit.jupiter.Testcontainers;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
 @SpringBootTest
-public class JdbcTgChatServiceTest extends AbstractTgChatServiceTest {
-    @DynamicPropertySource
-    static void jdbcProperties(DynamicPropertyRegistry registry) {
-        registry.add("app.database-access-type", () -> "jdbc");
+@Transactional
+@Rollback
+@Testcontainers
+public class AbstractTgChatServiceTest extends IntegrationTest {
+    @Autowired
+    private TgChatService tgChatService;
+
+    @Test
+    public void serviceShouldRegister() {
+        tgChatService.register(10L);
+        assertThatThrownBy(() -> tgChatService.register(10L)).isInstanceOf(ChatReregisteringException.class);
+    }
+
+    @Test
+    @Sql("/sql/insert-into-chats-table.sql")
+    public void serviceShouldUnregister() {
+        tgChatService.unregister(1L);
+        assertThatThrownBy(() -> tgChatService.unregister(10L)).isInstanceOf(ChatNotFoundException.class);
     }
 }
