+++ scrapper/src/test/java/edu/java/scrapper/domain/service/jpa/JpaLinkServiceTest.java
@@ -1,13 +1,17 @@
 package edu.java.scrapper.domain.service.jpa;
 
+import edu.java.scrapper.domain.service.AbstractLinkServiceTest;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.Rollback;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
 import org.springframework.transaction.annotation.Transactional;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
 @SpringBootTest
-@Transactional
-@Rollback
-@Testcontainers
-public class JpaLinkServiceTest {
+public class JpaLinkServiceTest extends AbstractLinkServiceTest {
+    @DynamicPropertySource
+    static void jpaProperties(DynamicPropertyRegistry registry) {
+        registry.add("app.database-access-type", () -> "jpa");
+    }
 }
