+++ scrapper/src/test/java/edu/java/scrapper/domain/service/jpa/JpaTgChatServiceTest.java
@@ -1,4 +1,12 @@
 package edu.java.scrapper.domain.service.jpa;
 
-public class JpaTgChatServiceTest {
+import edu.java.scrapper.domain.service.AbstractTgChatServiceTest;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+
+public class JpaTgChatServiceTest extends AbstractTgChatServiceTest {
+    @DynamicPropertySource
+    static void jpaProperties(DynamicPropertyRegistry registry) {
+        registry.add("app.database-access-type", () -> "jpa");
+    }
 }
